+++ src/test/groovy/domain/MCV.groovy
@@ -1,4 +1,4 @@
-package main
+package domain
 
 class MCV {
