+++ src/test/groovy/domain/Minigunner.groovy
@@ -1,4 +1,4 @@
-package main
+package domain
 
 class Minigunner {
