+++ src/test/groovy/domain/Point.groovy
@@ -1,4 +1,4 @@
-package main
+package domain
 
 class Point {
     int x
