+++ src/test/groovy/domain/ResetOptions.groovy
@@ -1,4 +1,4 @@
-package main
+package domain
 
 class ResetOptions {
     boolean drawShroud
