+++ src/test/groovy/domain/Sandbag.groovy
@@ -1,4 +1,4 @@
-package main
+package domain
 
 class Sandbag {
