+++ src/test/groovy/main/ShroudTests.groovy
@@ -1,31 +1,17 @@
 package main
 
-import client.MikeAndConquerGameClient
-import spock.lang.Specification
+
+import domain.Minigunner
+import domain.MovementDirection
+import domain.Point
+import domain.ResetOptions
+import util.ImageUtil
 
 import javax.imageio.ImageIO
 import java.awt.image.BufferedImage
 
 
 class ShroudTests extends MikeAndConquerTestBase {
-
-//    MikeAndConquerGameClient gameClient
-//
-//    def setup() {
-//        String localhost = "localhost"
-//        String remoteHost = "192.168.0.146"
-////        String host = localhost
-//        String host = remoteHost
-//
-//        int port = 11369
-//        boolean useTimeouts = true
-////        boolean useTimeouts = false
-//        gameClient = new MikeAndConquerGameClient(host, port, useTimeouts )
-//        gameClient.resetGame()
-//        gameClient.leftClickInWorldCoordinates(1,1)  // to get mouse clicks in default state
-//    }
-
-
 
     def "Shroud screenshot scenario 1"() {
         given:
