+++ src/test/groovy/util/ImageUtil.groovy
@@ -1,4 +1,4 @@
-package main
+package util
 
 import java.awt.image.BufferedImage
